types

+++ esphome/components/remote_base/electra_protocol.h
@@ -11,9 +11,9 @@ struct ElectraData {
   {
       uint64_t value;
       struct __attribute__((packed)) {
-            uint64_t padding : 36;
-            uint8_t temperature : 4;
-            uint32_t something : 24;
+            unsigned int padding : 36;
+            unsigned int temperature : 4;
+            unsigned int something : 24;
       } fields;
   } payload;
